Refactoring single file into multiple files to make it a bit more manageable. Exposing variables that can be configured in docker-compose.yaml such as DB_HOST, REDIS_HOST and POOL_SIZE Changing network mode to host to avoid the bottlenecks in bridge mode
