+++ nginx.conf
@@ -7,9 +7,11 @@ events {
 }
 http {
     access_log off;
+    error_log /dev/null emerg;
+
     upstream api {
-        server api01;
-        server api02;
+        server localhost:8080;
+        server localhost:8081;
         keepalive 200;
     }
     server {
